refactor(openomni): split ingress authority responsibilities
